+++ tests/provider/macros.rs
@@ -584,7 +584,7 @@ macro_rules! generate_language_model_tool_tests {
 
             // define tool function body, should return Result<String, String>
             #[allow(unused_variables)]
-            let func = ToolExecute::new(Box::new(|inp: Value| {
+            let func = ToolExecute::from_sync(|_ctx, inp: Value| {
                 // Ai SDK will pass in a json object with the following structure
                 // ```json
                 // {
@@ -593,7 +593,7 @@ macro_rules! generate_language_model_tool_tests {
                 // ```
                 let location = inp.get("location").unwrap();
                 Ok(format!("Cloudy"))
-            }));
+            });
 
             // define tool input structure
             #[derive(schemars::JsonSchema, Debug)]
@@ -782,10 +782,10 @@ macro_rules! generate_language_model_hook_tests {
 
             #[tool]
             fn get_neighbourhood() -> Tool {
-                Ok("".to_string())
+                Ok("ankocha".to_string())
             }
 
-            let _ = LanguageModelRequest::builder()
+            let result = LanguageModelRequest::builder()
                 .model($tool_model)
                 .system("Call the tool. Return the neighborhood. Nothing more and nothing less")
                 .prompt("What is the neighborhood?")
@@ -800,6 +800,8 @@ macro_rules! generate_language_model_hook_tests {
                 .generate_text()
                 .await
                 .unwrap();
+
+            assert!(result.text().is_some());
 
             let log = log.lock().unwrap();
             // Check pairs of prepare/finish